ALTERED RPM BASED APPS TO REQUIRE repos::all INSTEAD OF repos
previously, repos::all was called by base and then used by the apps but
the apps only required repos.
this meant that sometimes the class called the app before base had a
chance to install it.
this show that requiring a class will not bring in the classes whole subtree.