CREATE_RESOURCES NOW WORKING WITH HIERA HASH
after much consternation. compilation errors were being thrown because of a hard tab in the hiera data. which was obviously invisible. error

Error: malformed format string - %S at /tmp/vagrant-puppet-1/modules-0/users/manifests/accounts.pp:3 on node

helpful - considering the error states the string is in the manifest.pp when it is in fact, in the hiera file.

TODO * refactor the module to be the right way around. init.pp should call the subclass, not the other way around.