UPDATE MAN MODULE. USED .fixtures.yml & LOCAL GEM
use .fixtures.yml file to auto load the stdlib module when running test. this negates the need to clone it from github.com each time the test is run.

alter Gemfile to find the rspec-puppet gem locally instead of trying to clone it from the interwebs on first use. currently pointing to a .bundler path.

TODO git clone the rspec-puppet gem into somewhere suitable and point the gemfile there instead of the .bundler folder. or maybe not?? - see if the .bundler folder dynamially changes on update.