FIX CYCLICAL DEPENDENCY ERRORS IN REPOS CLASS
putting the purge first ensures that the auto-required folder is in place before placing the
files that auto-require it. the purge can take place before the files are placed because
puppet knows that it manages the file before it actually places it.

TODO create a proper rspec fixture to enable the tests to find the hiera files