COMPLETELY WORKING PROTOTYPE F/W TYPE & PROVIDER
for each linux_firewall resource instance, the provider iterates over
each line of the output of "iptables -S", regex checking for a match
against the resource instance's source, dest and commment. (this regex
will need to be updated when support for ports and protocols is
implemented). if a match is found, the variable "finding" is set to 1,
else it remains nil.

a proc was chosen over a block to allow the setting of the "finding"
varible outside of the proc and allow the variable to be set to 1 if
criteria are met anywhere during the iteration within the proc.

a proc was chosen over a lambda because of the way the return command
is handled by procs.

having said that, as iteration over resource instances is not required
it may be possible to use a simple block.

http://awaxman11.github.io/blog/2013/08/05/what-is-the-difference-between-a-block/

TODO

revert source/dest back to properties
add ordering logic
tests!!!!