REFACTOR TEST TO GIVE AN EXPECTED ERROR MESSAGE
this error message is much better (or just different) because it
basically says that ruby can't find the iptables binary within its
path. so if this test was run on a linux box that could find iptables
in its path, the test would pass..
right?

error;

  Puppet::Error:
    Command iptables is missing

putting a dummy executable file called 'iptables' somewhere in the path
alters (fixes?) the error to a "more correct" 'unable to match regex'
type of error.

changes;

* put the class declaration into a variable called 'provider_class'.

* use a 'before :each' block inside a context instead of a 'let' block
  outside of a describe block.

* use a 'context' block instead of a 'describe' block.

* use '@provider = new linux_firewall' instead of 'let (:provider )'

* call '@provider.destroy.should' instead of 'provider.destroy.should'